Refactoring: formating and name changes

+++ src/definitions.h
@@ -5,12 +5,12 @@
 
 #define VERSION "2.0.2"
 
-const int DEFAULT_ORDER{4};
+const size_t DEFAULT_ORDER{ 4 };
 
 // Minimum order is necessarily 3.  We set the maximum
 // order arbitrarily.  You may change the maximum order.
-const int MIN_ORDER{DEFAULT_ORDER - 1};
-const int MAX_ORDER{20};
+const size_t MIN_ORDER{ DEFAULT_ORDER - 1 };
+const size_t MAX_ORDER{ 20 };
 
 using KeyType = int64_t;
 using ValueType = int64_t;
